[dawn][vk] Finish MapMemoryAndPerformOperation spanification
Make functors use Span instead of std::span, use CopyFrom to fix a
DAWN_UNSAFE_TODO, and change an argument to be a size_t since mappable
buffers fit in the CPU address space.

Bug: 439062058
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/334956
Commit-Queue: Corentin Wallez <cwallez@chromium.org>

+++ src/dawn/native/vulkan/BufferVk.h
@@ -106,7 +106,7 @@ class Buffer final : public BufferBase {
     // unmapped/flushed if necessary. The op function receives a span of exactly the requested
     // size.
     template <typename F>
-    MaybeError MapMemoryAndPerformOperation(uint64_t requestedOffset, size_t requestedSize, F&& op);
+    MaybeError MapMemoryAndPerformOperation(size_t requestedOffset, size_t requestedSize, F&& op);
 
     MaybeError MapAsyncImpl(wgpu::MapMode mode, size_t offset, size_t size) override;
     MaybeError FinalizeMapImpl(BufferState newState) override;
